Create CODEOWNERS (#14)
The CODEOWNERS file has been created to protect the documentation related directories to ensure that any changes to those directories are reviewed by people on the Framework team.  As part of this I have also created the @extremenetworks/autoiq-framework team and provided Write access to the repositories to members of that team.