CI fixes (#117)
* Fix testcase filename check

The main problem was the carrot(^) at the beginning of the regex. It was causing the regex to look for "test_" at the beginning of the filename string. The filename string contains all of the folder structure still, so this was failing.

Also, made the regex a bit more specific to try to avoid false positives.

* Exempting init.py from directory check