Update topo.prod.2degrees.yaml (#756)
Changing the "sw_capwap_url" as hac.extremecloudiq.com instead of 2degrees.extremecloudiq.com as currently in our VSP device in Blr-tb-2 the VOSS device is not accepting the RDC url as "2degrees.extremecloudiq.com"- initiating a mail thread with the Wired ST team to check out if its a known issue/limitation. In the meantime or as always expected in production by default the switches will have(if it was reset) configured with URL for XIQ as "hac.extremecloudiq.com". Hence modifying the URL for this 2degrees accordingly.