Update CODEOWNERS (#761)
* Update CODEOWNERS

Ensuring that we have the XIQ environments used for Production Sanity suite are changed with proper review - made Banu and Philip as owners along with AutoIQ-framework team (though i want Banu or Philip to be mandatory reviewers and dont want changes to happen here without them knowing about it).

CC @shreenbanu  @pdungdo

* Update CODEOWNERS

Made correction to the username for Philip. Thanks @shreenbanu for pointing it out.