2022-12-20 - Digital Twin Sanity Test Update (#1086)
The Digital Twin Sanity test is run via AutoIQ.  The test failed this AM due to the Device360 view already being closed for the Digital Twin device.  The issue happened during the "Teardown" section of the Test Case with the code "Close Device360 window".

I had included this keyword within the "Teardown" section in case there was an error while obtaining data from the Device360 view.  I did not want the view to remain open.

With the changes to add IRV to Device360.py file (PR-758) when the "Close Device360 window" keyword is run during the "Teardown" section, the call fails since the Device360 view was already successfully closed with the "Get Switch Information" keyword.  Prior to the changes, the keyword would "silently" fail and not result in the Test Case failing.

I have removed the "Close Device360 Window" keyword from the "Teardown" section.

- DigitalTwin_Sanity_XIQ7860.robot
  -- TCCS-13499: removed "Close Device360 Window" from "Teardown" code.